added labels in property image upload apis

+++ app/views/properties.py
@@ -176,6 +176,7 @@ class PropertyUpdateView(MethodView):
 
             data = request.form 
             files = request.files.get('image')
+            label = data.get('label')
             
             if not data and not files:
                 return jsonify({'error': 'No data in payload'})
@@ -225,6 +226,9 @@ class PropertyUpdateView(MethodView):
             current_app.db.properties.update_one({'_id': ObjectId(property_id)}, {'$set': update_data})
             # Add image if provided
             if 'image' in request.files:
+                if not label:
+                    return jsonify({'error': 'Missing image label'})
+
                 file = request.files['image']
                 org_filename = secure_filename(file.filename)
                 timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
@@ -235,13 +239,14 @@ class PropertyUpdateView(MethodView):
                 if os.path.exists(os.path.join(user_media_dir, filename)):
                     return jsonify({'error': 'File with the same name already exists in the folder'}), 200
                 # Check if image with same name already exists in database
-                if any(org_filename == image_name for image_name in [image.split('/')[-1] for image in property_data.get('images', [])]):
+                if any(org_filename == image_name for image_name in [image.get('name') for image in property_data.get('images', [])]):
                     return jsonify({'error': 'File with the same name already exists in the database'}), 200
                 
                 image_path = os.path.join(user_media_dir, filename)
                 file.save(image_path)
                 image_url = url_for('serve_media', filename=os.path.join('user_properties', str(user['uuid']), str(property_id), filename))
-                update_data.setdefault('images', []).append(image_url)
+                image_data = {'label':label, 'name': filename, 'image_url':image_url}
+                update_data.setdefault('images', []).append(image_data)
                 current_app.db.properties.update_one(
                     {'_id': ObjectId(property_id)},
                     {'$push': {'images': update_data['images'][0]}}
@@ -252,6 +257,7 @@ class PropertyUpdateView(MethodView):
             return jsonify({'message': 'Property information updated successfully'}), 200
         else:
             return jsonify({'error': 'User not found'}), 200
+
 
 class PanoramicImageView(MethodView):
     decorators = [custom_jwt_required()]
@@ -468,6 +474,7 @@ class PanoramicImageView(MethodView):
         log_action(user['uuid'], user['role'], "deleted-panoramic-images", {"property_id": property_id, "property_version": property_version, "order": order})
         return jsonify({"message": "Panoramic image deleted successfully"}), 200
 
+
 class PropertyImageDeleteView(MethodView): 
     decorators = [custom_jwt_required()]
     
@@ -481,55 +488,56 @@ class PropertyImageDeleteView(MethodView):
         except EmailNotValidError:
             user = current_app.db.users.find_one({'uuid': current_user})
 
-        if user:
-            if user.get('role') == 'realtor':
-                return jsonify({'error': 'Unauthorized access'}), 200
-            
-            data = request.json
-            property_id = data.get('property_id')
-            image_url = data.get('image_url')
-            
-            if not property_id or not image_url:
-                return jsonify({'error': 'property_id or image_name is missing in the request body'}), 200
+        if not user:
+            return jsonify({'error': 'User not found'})
 
-            property_data = current_app.db.properties.find_one({'_id': ObjectId(property_id)})
-            property_seller_data = current_app.db.property_seller_transaction.find_one({'property_id': property_id, 'seller_id': user['uuid']})
-            
-            if property_data is None or property_seller_data is None:
-                return jsonify({'error': 'Property does not Exists or you are not allowed to delete image to this property'}), 200
+        if user.get('role') == 'realtor':
+            return jsonify({'error': 'Unauthorized access'})
 
-            if not property_data:
-                return jsonify({'error': 'Property not found'}), 200
-            
-            file_name = os.path.basename(image_url)
-           
-            # Delete the file from the server
-            user_media_dir = os.path.join(current_app.config['UPLOAD_FOLDER'], 'user_properties', str(user['uuid']), str(property_id))
-            file_path = os.path.join(user_media_dir, file_name)
+        data = request.json
+        property_id = data.get('property_id')
+        image_url = data.get('image_url')
+        label = data.get('label')
 
-            if not os.path.exists(file_path):
-                return jsonify({"error":"file does not exits "})
-            
-            property_images = property_data.get('images')
-            if image_url not in property_images:
-                return jsonify({"error":"file does not exits "})
-            
-            os.remove(file_path)
+        if not property_id or not image_url or not label:
+            return jsonify({'error': 'property_id, image_url, or label is missing in the request body'})
 
-           # Remove the image URL from property_data's images list
-            property_data['images'].remove(image_url)
+        property_data = current_app.db.properties.find_one({'_id': ObjectId(property_id)})
+        property_seller_data = current_app.db.property_seller_transaction.find_one({'property_id': property_id, 'seller_id': user['uuid']})
 
-            # Update the user's properties in the database to reflect the removed image
-            result = current_app.db.properties.update_one(
-                {'_id': ObjectId(property_id)},
-                {'$set': {'images': property_data['images']}}
-            )
+        if not property_data or not property_seller_data:
+            return jsonify({'error': 'Property does not exist or you are not allowed to delete image from this property'})
 
-            if result.modified_count == 0:
-                return jsonify({'error': 'Failed to delete the image'}), 200
-            
-           
-            log_action(user['uuid'], user['role'],"deleted-property-image", data)
-            return jsonify({'message': 'Image deleted successfully'}), 200
-        else:
-            return jsonify({'error': 'User not found'}), 200
+        file_name = os.path.basename(image_url)
+        user_media_dir = os.path.join(current_app.config['UPLOAD_FOLDER'], 'user_properties', str(user['uuid']), str(property_id))
+        file_path = os.path.join(user_media_dir, file_name)
+
+        print(file_path)
+
+        if not os.path.exists(file_path):
+            return jsonify({"error": "File does not exist"})
+
+        os.remove(file_path)
+
+        property_images = property_data.get('images', [])
+
+        # Find the image object to be removed
+        image_to_remove = next((image for image in property_images if image['image_url'] == image_url and image['label'] == label), None)
+
+        if not image_to_remove:
+            return jsonify({"error": "Image with the specified URL and label does not exist"})
+
+        # Remove the image object from the property_data's images list
+        property_images.remove(image_to_remove)
+
+        # Update the user's properties in the database to reflect the removed image
+        result = current_app.db.properties.update_one(
+            {'_id': ObjectId(property_id)},
+            {'$set': {'images': property_images}}
+        )
+
+        if result.modified_count == 0:
+            return jsonify({'error': 'Failed to delete the image'})
+
+        log_action(user['uuid'], user['role'], "deleted-property-image", data)
+        return jsonify({'message': 'Image deleted successfully'})
